Update ternlabel.m
Hi! first of all, thank you for this amazing ternary diagram creator, i`m a chemical engineering student and this is awesome. Nevertheless, I have a little concern about the label position.
The text are too close to the numbers of the diagram and it made mine look like this: https://pasteboard.co/I3oawRe.png
So it might be better to change the numbers a little, for me the following numbers worked better: https://pasteboard.co/I3ocsgQ.png
I´m just a begginer to programming, but i hope it will be useful.